+++ lib/eligible.rb
@@ -44,6 +44,7 @@ require 'eligible/icd'
 # New REST API Endpoints
 require 'eligible/v1_0/rest_api_base'
 require 'eligible/v1_0/charge'
+require 'eligible/v1_0/contract'
 require 'eligible/v1_0/enrollment'
 require 'eligible/v1_0/estimate'
 require 'eligible/v1_0/estimate_service_line'